// src/features/students/students.schema.ts
import z from 'zod'

export const studentSchema = z.object({
  id: z.string().optional(),
  username: z
    .string()
    .min(3, { message: 'Username must be at least 3 characters long!' })
    .max(20, { message: 'Username must be at most 20 characters long!' }),
  password: z
    .string()
    .min(8, { message: 'Password must be at least 8 characters long!' })
    .or(z.literal('')),
  name: z.string().min(1, { message: 'First name is required!' }),
  surname: z.string().min(1, { message: 'Last name is required!' }),
  email: z
    .string()
    .email({ message: 'Invalid email address!' })
    .optional()
    .or(z.literal('')),
  phone: z.string().optional(),
  address: z.string(),
  img: z.string().nullish(),
  bloodType: z.string().min(1, { message: 'Blood Type is required!' }),
  birthday: z.string().min(1, { message: 'Birthday is required!' }),
  sex: z.enum(['MALE', 'FEMALE'], { message: 'Sex is required!' }),
  gradeId: z.number().min(1, { message: 'Grade is required' }),
  classId: z.number().min(1, { message: 'Class is required' }),
  parentId: z.string().min(1, { message: 'Parent Id is required' }),
})

export type StudentFormValues = z.input<typeof studentSchema>
export type StudentSchema = z.output<typeof studentSchema>